+++ app/src/main/java/com/vci/vectorcamapp/imaging/presentation/ImagingAction.kt
@@ -3,7 +3,7 @@ package com.vci.vectorcamapp.imaging.presentation
 import android.graphics.Bitmap
 import com.vci.vectorcamapp.core.domain.util.Result
 import com.vci.vectorcamapp.core.domain.util.imaging.ImagingError
-import com.vci.vectorcamapp.imaging.domain.BoundingBoxUi
+import com.vci.vectorcamapp.imaging.presentation.model.BoundingBoxUi
 
 sealed interface ImagingAction {
     data class UpdateSpecimenId(val specimenId: String) : ImagingAction
